Limited support for making HTTP requests
This commit adds the basic architecture of hydra. There is a `Hydra`
type which is a proxy for a set of `Worker`s. Each worker runs an event
loop (from the _mio_ package). Each event loop can support a large
number of connections. A `Connection` is tied to this event loop (and
hence a worker). Each Connection owns an `Http2` state machine
(_solicit_ is used as the protocol implementation). The HTTP/2 protocol
enables multiplexing on the connection level, so each of these state
machines may have a number of streams.

To summarize, the top down architecture looks like this:

* `Hydra` - Proxy to set of N workers
* `Worker` - A thread with an event loop multiplexing `Connection`
* `Connection` - A TcpStream + Http2 state machine; multiplexes requests
* `Request` - an HTTP request with a handle. Runs on a connection.

Workers can be considered an implementation detail of hydra. However, a
consumer must consider the Hydra proxy, connections, and requests.
Hydra, because it's the main entry point of the client, connections due
to potentially stateful nature of them, and requests because those are
highly application specific.