Enable customizable assertions in live tests
Up until now, the collector would make a few assertions on the returned
response. However, this wasn't going to scale to requests other than the
initially very basic one we were making.

To support this, a CheckResponse trait was added. It has a single
method, `check`, which should run some assertions. A type implementing
CheckResponse can be passed to the collector's new `check_responses`
method. The method iterates over all responses the collector received
and runs the response checker.

Implementing this manually for different requests would have turned
into a lot of boilerplate. A macro `response_spec!` was added to address
this. The macro lets a developer think in terms of http attributes like
headers, the status, and what should be in the body. The tests added in
this commit demonstrate the usage.