Propagate remaining unhandled result errors
All* errors are now propagated up to a client level or the top of an
event loop (in the case of errors in a non-user-accessible thread). In
the case of errors in the event loop, the connection is deregistered and
thrown out. It's up to the protocol layer to make sure all the streams
are handled first. User facing errors are naturally left to the user to
handle.

This marks the first time there are no compiler warnings!

* There are certain errors not handled yet that don't show up in the
  form of a result such as an error event being returned from mio.